+++ titiler/main.py
@@ -3,12 +3,12 @@
 from . import settings, version
 from .endpoints import cog, mosaic, stac, tms
 from .errors import DEFAULT_STATUS_CODES, add_exception_handlers
-from .middleware import CacheControlMiddleware, TotalTimeMiddleware
+from .middleware import (CacheControlMiddleware, CompressMiddleware,
+                         TotalTimeMiddleware)
 
 from fastapi import FastAPI
 
 from starlette.middleware.cors import CORSMiddleware
-from starlette.middleware.gzip import GZipMiddleware
 
 api_settings = settings.ApiSettings()
 
@@ -35,7 +35,7 @@ if api_settings.cors_origins:
         allow_headers=["*"],
     )
 
-app.add_middleware(GZipMiddleware, minimum_size=0)
+app.add_middleware(CompressMiddleware, minimum_size=0)
 app.add_middleware(CacheControlMiddleware, cachecontrol=api_settings.cachecontrol)
 app.add_middleware(TotalTimeMiddleware)
